Add halo update for calvingVelocity in von_Mises_calving
Results from von Mises calving were not bit-for-bit across
different numbers of processors. This fixes that issue. It is not
apparent why this is necessary here, and why it is not necessary
for eigencalving, so perhaps the issue is rooted in the fields
used to calculate calvingVelocity in the von_Mises_calving routine.