SCREAM: specify SCREAM_TEST_DATA_DIR when building baselines
This is *usually* not necessary, since it defaults to
CMAKE_BINARY_DIR/data, which is fine. However, if one has
this folder specified in a mach file to some defaults location,
running test-all-scream -b <ref> would cause havoc, with the
actual build happening in ctest-build/baselines/<build>,
but SCREAM_TEST_DATA_DIR still set to the machine default
location.