SCREAM: fix undesired behavior of --quick-rerun in test-all-scream
The idea is not to skip the config stage, but rather to not remove
the cmake files (e.g., the cache). If cmake runs but nothing changes
in the cache or in the generated files, make will not rebuild the
whole project anyways. However, if we are in the middle of development,
and the user changes a cmake file, we must make sure cmake is re-run.