SCREAM: change how we communicate restarted run to the AD
For when the 'initialize' method is called (standalone scream runs),
we still check the input parameter list. But when we call the
individual initialization methods, we can pass the run type to
initialize_fields, which is where the AD needs to know whether
this is an initial or restarted run.