Force ISMIP6-retreat i/o on restart to initialize TF properly
This commit replace the fix of the previous commit with a solution that
also solves the restart BFB error for the ISMIP6 retreat param. when the
restart time does not match a time in the ISMIP6 forcing file.  It does
so by always forcing the previous forcing data to be read on the initial
timestep of a restart run.  This would be a redundant read if the
initial time of the restart happens to be a time in the forcing file,
but is necessary otherwise.  Rather than add logic to figure that out, I
just have that extra read always happen on the initial time of a restart
run.  That has the benefit of making the initial timestep of a restart
and a cold start use the same logic in the ISMIP6 retreat routine.

Before this commit, the new
landice/humboldt/mesh-3km_restart_test/velo-none_calving-ismip6_retreat
test failed the validation step.  After this commit, that test passes.