Add virutal temperature diagnostic, add files for other diagnostics
This commit adds the virtual temperature diagnostc and adds the files
for the other physics functions diagnostics we want to create.  Note, the
other diagnostics haven't been developed yet, at this commit there are the files
and nothing else.

With the exception of vertical_height_interface, which will follow a different
pattern to the others since interface height is a column-wise operation instead
of a scalar operation.