Add calving CFL calculation to other calving laws
This commit adds the calving CFL calculation to all the other calving
laws that use a calvingVelocity: specified_velocity, eigencalving,
ISMIP6-retreat, and damage-rate
This calculation should be low cost, so I've enabled it in these
routines, regardless of if it is being used.  I anticipate its result
will be a useful diagnostic even if it is not being applied in the
timestepper.