Calculate calving on initial time when not a restart
This is needed to get the calving CFL to work properly on the first
timestep because the calving CFL calculation is made on the timestep
prior to its application.  To ensure there is no actual calving for
rate-based calving laws, the model deltat is temporarily set to 0 before
calling the calving routines.  Note that non-rate-based calving laws
(thickness, damage thresholds, etc.) will now be applied on the
initial time whereas before they were not.  But I think that behavior is
actually preferable.  We don't have any tests or routine configurations
using those calving laws currently, so that change will not disrupt
anything.