Fixes to dx_from_area unit test and to shoc_standalone
This commit fixes the syntax in the calculate_dx_from_area unit
tests to accomodate single-precision builds.

It also fixes the input file for shoc_standalone to load the latitude
which is now needed to back out the cell_length.  Before latitude was
NaN for these tests which threw a FPE.