Fix bug in SPA when running simulations with levels not = 72
This commit fixes a bug in the SPA interface where it was assumed that
the SPA data and the simulation would have the same number of vertical
levels.  Since SPA data is on a 72 level grid this meant we could not
run at 128 levels.

This commit fixes the definition of how many levels are in the SPA data
to only come from reading that info from the SPA data file.  Presumably
this would also allow us to use SPA data with a number of levels different
than the current 72.

Fixes #1646