Fix to p3 interface and code to accomodate cases where nlev % packsize = 0
This commit increases the size of the workspace manager in the p3 interface
to use nk_pack_p1 which is the number of packs for num_levels + 1.

It also fixes the sedimentation code for rain, cloud and ice to use
the numver of packs rather than the extent of the variables for a Kokkos
loop.  This is necessary becasue the tendency variables will have a larger
extent given the nk_pack_p1.