Switch init condition file for all driver based tests to use nccn
This commit switches the name of the nc_activated variable in all of the
driver based tests to use nccn.

This commit also simplifies the driver based tests to all use the same
initial condition file, so in the future there would only need to be
one file to change if we change the name of a variable.