Fixed use_case files for SSP370 and SSP585
The EAM use_case files for SSP370 and SSP585 contained parameter settings
meant for EAMv1 that need to be cleared. surfdata_map files for ELM are
updated to use one for simyr2015 that is consistent with the starting year
of the SSP scenarios. landuse timeseries file for SSP585 is corrected with
the proper ne30pg2 grid as default. Several consistency check flags that
are typically used for hybrid run are specified as default.

Tests for the SSP370 and SSP585 compsets are introduced.

The tests, the update for the ELM SSP585 use_case file for maint-2.0 went in
via PR #4887.

The initial SSP370 and SSP585 use_case files went into master via PR #4844.

[BFB] except for SSP370 and SSP585. wcprodssp tests are new.