Revert "Modify flood fill to allow it to optionally detect hinges"
This reverts commit 6598c5e91274dd5ed30f4b705045092d7fd78044.

Reverting previous commit.  After testing, it is questionable how
valuable the hinge removal implemented here will be.  For one thing,
Albany is using a slightly different mesh, because it includes the
extended row of cells.  So the hinges detected here might not be the
same as hinges on the actual Albany mesh.  Secondly, the hinges here
might be overly aggressive, in that any time an ice shelf grows forward,
there are likely to be some cells only attached to one other cell, and
those would get removed before the surrounding cells even have a chance
to fill in.  For those reasons I am removing the hinge algorithm so as
not to clutter the code.  But I'm using a revert rather than removing
the commit so this algorithm is saved in the git history if we are need
to resurrect it for some reason.