This commit removes unneeded default output for our EAMxx CIME runs.
The current set of default output for CIME runs includes everything
produced by all processes.  But many of these are not used for post
processing and analysis and thus don't need to be written by default.

A user can always include them before simulation if they are needed
for debugging.