Rebase and cleanup a bit.
This commit rebases the branch on the current master since the
branch was getting stale and cleans up the tests to match recent
changes.

This commit also fixes the atm_density diagnostic following recent
changes to the atmosphere_diagnostic class.

Finally, this commit is incomplete, there are still failing tests that
need to be resolved in a subsequent commit.