Changes to AD to query parameter list for casename.
This commit causes the AD to query the parameter list for the E3SM casename
and assign it as the AD casename.

This commit passes the casename to the output manager if an output control file
doesn't already specify a casename for the file.  There is a running tally that
ensures that if multiple output streams are requested with the default filename
that each has a unique full filename, in particular the files will use the *.hX.*nc
convention where X is an integer counter for each file.