Fix diagnostics missing team_barrier()
This commit fixes the dry_static_energy, vertical_layer_interface,
and vertical_layer_midpoint diagnostics which were failing the tests.

The solution was to include team.team_barrier() in the Kokkos for loops
in the diagnotic run_impl definition.

After this commit only the sea_level_pressure test is still failing.  That
will be fixed in the next commit.