Update the initial condition files used by the stand alone tests.
This commit updates the initial condition file used by stand alone
testing to include the dry pressure coordinates.

This commit also streamlines where that file is defined.  Up until now
each stand alone test explicitly defined the location and name of the initial
condition file in its own input.yaml file.  Although the file was the same
for all of the tests.  This meant a lot of repeated code and made it tedious
to update the file.

This commit adds a cmake variable to the tests that defines the name of the
initial condition file they all use, so when we update the file again we only
have to change the name in once location, namely `components/scream/tests/CMakeLists.txt`.