Move the caclulation of some preprocessed variables in the p3 preprocessing.
Currently the mass mixing ratios are converted from wet to dry in the preprocessing
after all other variables are calculated.  This has the risk of a calculation using
a wet variable when it should have been dry.  This commit simply changes the order of
operations so that MMR are converted first, before all other calculations.

An example of this would be in the calculation of `dz` which is dependent on the water
vapor `qv`.