Move refactored cospsimulator_intr.F90 to MMF dir
Move refactored and MMF-enabled cospsimulator_intr.F90 to MMF dir, and
revert to original version of cospsimulator_intr.F90 in cam dir. The
refactored version is compatible with non-MMF configurations, but does
require building crmdims.F90 to provide the CRM dimension sizes, which
is not currently built for non-MMF configurations.