EAMxx: change PropertyCheck::check return type
* The return type contains a CheckResult enum, which can be
Pass, Fail, Repairable. That means that PC can return whether
the failure was "bad" or "not-so-bad", so that we can decide
if we can repair it.

* The class FieldWithinIntervalCheck (and its derived ones) can
now store "repairable" bounds, which are no-tighter than the
original bounds. If the field has values outside the original
lb/ub values, but within the repairable lb/ub values, then
the check returns Repairable, and repair is allowed.

* Removed CheckAndRepairWrapper. The class was badly designed,
and did not deliver what it was supposed to deliever, b/c the
check and repair objects had no way to communicate. With the
addition of "repairable" lb/ub in FieldWithinIntervalCheck,
this wrapper is also no longer needed.