Adjust the io_test to support a generic function for test values.
This commit cleans up the code in io.cpp and supports generating
test data using the function generate_data_xy.  It also introduces
the function check_data_xy which works with generate_data_xy.

Now we should be able to simply change the generate_data_xy function
in the io_test to test simpler or more complex functions for data.