EAMxx: Always enable EKAT's FPE machinery, but use scream's default fpes
The EKAT machinery for FPEs is *always* needed, since even if we don't
want FPEs in scream, we still need to be able to disable FPEs enabled
by e3sm upon entry in the cxx code (and re-enable e3sm fpes before
returning control to the mct coupler).

The important thing is to pick the correct value for the default FPEs.
So always set EKAT_ENABLE_FPE=ON, but *never* use ekat's default fpes.