Add a new horizontal remap utility
This commit adds the structure for a new general horizontal remapping
utility.  This utility can be used to remap source data from fine to
coarse and coarse to fine.

The purpose of this utility is to allow for general horizontal remapping
to be available to the entire EAMxx code base.  Currently the only process
that uses horizontal remapping is SPA, but there are other applications.

This first commit introduces the infrastructure and a set of basic unit tests
for that structure.  Including a randomized test that generate a set of remapping
weights and for random source to target columns.

Next step will be to develop a simple routine for gathering remap weights from
a file.  Another next step will be to shift SPA to begin using the general utility.