Add a function to read remap data from file and assign it to remapper.
This commit adds a routine to the horizontal remapper to read remap data
from file and set up all of the remap segments.  This should be a simple
way to set up a remapper based entirely on remap pairs and weights calculated
offline.

This commit also adds to the horizontal remapper test to test his new
functionality.