Switch SPA to using remap information using horizontal remap function.
This commit switches SPA to rely entirely on the horizontal remap
function to gather the remapping information.  The old code is still
there, with a comment to remove in a subsequent commit.  This is done
becuase debugging isn't finished and it might be nice to go back
and check against what was done before.

SPA still uses in house code to apply the remap.  A subsequent commit
will switch SPA over to using the horizontal remap to apply the remap.