Add overloads to apply_remap on 2D and 3D views of data.
This commit extends the apply_remap function to also work when the
source data and remapped data are defined as a set of horizontal
slices over
  levels - 2D
  levels+bands - 3D

This is a necessary step to getting SPA to use apply remap everywhere.

Note, apply_remap can also be extended to accomodate 4+D variables
of data, but that is not done here as it isn't needed.