Some more fixes for GPU simulations.
Missed a deep_copy in both the horizontal_remap test and in the SPA
interface.

This commit also adjusts the SPA standalone test to compare 4 snapshots
of output data in a single file.  The purpose is that SPA stand alone
runs at a timestep that should cause it to remap in multiple months, so
we don't want to just check that the last month remapped appropriately, we
want to check multiple months of remapping.

Also, supressed an annoying print statement that would print for all nodes to
just print on the master rank.