Switch apply_remap to be done entirely on Host.
This commit causes the entirety of the apply_remap call and loops
to be conducted on Host and then coped back to device at the end.

This provides a significant ( order 20x ) speedup over the previos
approach.  There may be even better performance that can be gained
by doing everything on Device, that will be explored in a subsequent
commit or PR.

Note, the slowdown this remedies was likely due to excesive host/device
copies.