EAM stealth feature: revised handling of cam_in%cflx
In the released v2, the tracer mixing ratio tendencies caused by
surface fluxes (cam_in%clfx, but excluding the water vapor flux)
are calculated in tphysac, inside the subroutine clubb_surface.
These tendencies are applied after clubb_surface before
the dry removal of aerosols.

In this commit, an option is added to move (1) the calculation of the
cflx-induced tendencies and (2) the update of tracer mixing ratios
from the original place in tphysac
to tphysbc, right before the cloud macro-microphysics subcycles.

The new option can be turned on by setting the namelist variable
cflx_cpl_opt = 2. (Note that this new option does not affect water
vapor or tracers whose cam_in%cflx is zero when clubb_surface is called.)

Minor conflict with master resolved on 2023-02-17 in:
	components/eam/src/physics/cam/physpkg.F90