Hommexx/SL: (Re)fix a variable for lambda capture in two spots.
HIP can't handle the reference, but we still need it for the F90 dycore with
HORIZ_OPENMP enabled, since in that code path we enter all of our routines
within a threaded region and so cannot create new Kokkos::Serial views during
time stepping. (Recall Kokkos::Serial is necessary b/c Homme's threading is not
compatible with Kokkos::OpenMP.) Thus, if COMPOSE_PORT is not defined, take a
reference.