Update the pressure level tests to incorporate random locations.
This commit fixes a few things in the interface that were causing
 general fails in the testing.  It also updates the test for the
field at pressure level slice diagnostic to look at a simple static
value and to take a random pressure level for testing.  The pressure
field in the test is also now varying over the number of columns too.