Fix mistake in new diagnostic test, flag problematic diags for general output
This commit adds a fix to the new field_at_pressure_level diagnostic
test, the fix is in how the pressure level is determined given a top
and surface pressure bounds.

The fix also flags 3 unrelated diagnostics that seem to not work when used
as output for a simple standalone test.  Those will have to be addressed
in a separate PR.