Create option for restore_calving_front to allow retreat
This commit modifies the restore calving code to handle retreat and advance
from the initial calving front extent separately.  There is a new option
config_restore_calving_front_prevent_retreat that can be used to prevent
or allow retreat when config_restore_calving_front is true.
config_restore_calving_front always prevents advance.  The motivation for
this is that preventing retreat is really only desired in E3SM where the
coupler cannot yet support a changing ice shelf extent.  For many standalone
configurations we want to prevent advance, but there really is no reason to
prevent retreat, and in fact, it is probably desirable to have retreat
if melt or mask calving causes it to happen.  Also, where
restore_calving_front prevents advance, set calvingVelocity to the advection
velocity.  This is useful for calculating the ismip6 variable licalvf in
post-processing.