Allow mask calving to be applied in addition to physics- based calving schemes
This allows ISMIP6 Tier2b experiments to be run with vM calving.  This is implemented
by eliminating 'mask' as a calving option and creating a new namelist option
config_apply_calving_mask that operates independent of the choice of config_calving.
Now mask_calving is checked and called after the physics-based calving laws.