Add a test to the coarsening remapper for when the remap is large.
This commit adds a small test that checks for the case when the number
of remap triplets exceeds the size of the mesh.  This is typical in
remap files that map between two meshes, mainly because every target
column is typically a summation of O(10) source columns.

In particular, when checking ne4->ne2 mapping it was noticed that the
remap file had >866 remap triplets and that the current test was
insufficient.

This commit also organizes the code of the original test to allow
for code reuse in both tests.  Other than that, the original test
is left unchanged.