Adjust the `get_field` function in scorpio output to take fm as input.
This commit adjusts the get_field call in the scorpio output class to
accomodate more than one field manager being stored in the output
class.  This is needed for horizontal remapping of diagnostics.

The old function used just the io_fm which was set by the remapper as
the only field manager tracked by the output class.  In our current
remapping approach we need to first update the diagnostics on the main
simulation field manager, then remap that diagnostic which happens
onto the io_field manager.  The get_field function only retrieved fields
from whatever the main io field manager was, which was on the io grid.
This made it difficult to update the diagnostics using the simulation
field manager.