Add checker to updating diagnostics to avoid updating > 1 time.
This commit responds to reviewer comments, getting rid of redundant code.

The main fix is to add a tracker that will log when a diagnostic has been
updated and skip updating it a second time, like in the case where we have
output that is a horizontal slice of a diagnostic, which under-the-hood is
a diagnostic of a diagnostic.