Update elm-fates to enable use of lightning data
This imports and adapts the clm-fates fire base types and
factory method of fire type creation.  The standard elm
FireMod has not been changed, but could adopt this method
in the future if so desired.  FATES will now have access
to the same lightning and population data the the FireMod
make use of.  The code also includes the hooks to enable
the use of additonal FATES fire modes, but these require
the use of data sets not currently stored on most machines.