Add a unit test to property check for single out of interval fails.
This commit will add a unit test that should catch a current bug in
the within interval field property checks when the maximum value is
violated.

The new test checks that the failed location matches the appropriate
column for both a violation of the minimum value and the maximum value,
two separate tests.