Add a mask to the field_at_pressure_level diagnostic
This commit adds, as an extra_data to the field_at_pressure_level
field, a mask field which can be used to flag locations where
data was extrapolated.  This will be helpful in ensuring that when
a field_at_pressure_level diagnostic is computed along w/ horizontal
remapping that masked values won't contribute to the overall values.