Apply mask inside the coarsening remapper to zero out masked values.
This commit applies the mask, where specified, to the local mat vec
function inside the coarsening remapper.  This effectively zeros
out any contributions from "masked" data.

The next step, and next commit, is to rescale the remapped data
by the missing weights, which should now be stored in the remapped
horizontal mask