Limit velocity throughout ice column
Use the entire vertical velocity profile to determine the scaling
factor when applying the upper limit on velocity. This is necessary
because whatever instability is causing unrealistic velocities (at
least in the ISMIP6 4km expAE02 case) is causing an inverted velocity
profile, i.e., basal speed is higher than surface speed, which should
not be the case. This uses the maximum velocity within the ice column
to determine the scaling factor to apply to limit the velocity
to config_unrealistic_velocity.