Adjust how velo cap is applied to normalVelocity
This commit changes the way the velocity limit is applied to
normalVelocity to account for the fact that one or both cells
adjacent to an edge might be modified and the magnitude to which
changes in cell velocities impact an edge velocity is a function
of the cell velocities.

To do this, a scaling_factor vector of length nCells is first
calculated.  Then the adjustment is applied to normalVelocity as a
function of the scaling_factor of its two neighbors.  Finally, the cell
center velocities are adjusted by the scaling factor.

This is probably overkill for a functionality that is pretty arbitrary
to being with (capping a velocity).  A simpler solution would be to just
apply the velocity limit to normalVelocity itself.  If that was done,
normalVelocity and uReconstructX/Y would no longer be consistent, but
that might not be a concern for a velocity field that was monkeyed with.