Add robust checking of the vertical remapper.
This commit adds a vertical_remapper test similar to the coarsening
remapper test.

This commit also reorganizes some of the vertical remapper and scorpio_output
code to accomodate better error checking.

There is a bugfix where the target interpolated field could have ILEV as a tag,
which in the case of vertical remapping doesn't make sense.  The vertical remapper
always remaps onto the same levels which so the level TAG should be the same for
all fields.

Finally, this commit puts in place the infrastructure to support vertical
remapping on fields of packsize != SCREAM_PACK_SIZE.  That being said, it doesn't
work yet.  This is because the target pressure levels are themselves a field which
is initialized with SCREAM_PACK_SIZE.  So when grabbing a view using a different packsize
the get_view() command throws an error.  This will need to be fixed to support
multiple packsizes, there is a TODO comment pointing this out.