Fix vertical interpolator to allow data and pressure levels of different packs.
This commit change the vertical interpolator to allow the source and target
pressure profiles to have been initialized with a different pack size than the
source and target data.  This expands the applicability of the vertical remapping
functions and the vertical remapper to accomodate fields which may have been allocated
with a different packsize than SCREAM_PACK_SIZE.

The simplest example would be any field that was initialized with packsize = 1.