Add a getter to IO interface to grab the timestamp of a file.
This commit introduces the function read_time_at_index to the IO
interface for EAMxx.  The purpose of this routine is to read the
specific "time" variable to get the timestamp from a file.

This commit also adds the read_time_at_index calls to the io_test
unit test which verifies that the output variables store the correct
timestamp and verifies that the read_time_at_index routine is working
correctly.