Add support for geo data in vertical remapper.
This commit updates the vertical remapper to provide support for
geo data in the vertical interpolation target grid.  This commit
also updates the mask default value to be compatible with single
precision output.

Finally, this commit updates how masks and mask arrays in general
are supported in both vertical and horizontal remapping.