Homme: Fix bubble_init's initialization of dp3d.
Previously, it was computing it according to the analytical formula. But that is
not BFB in finite precision with the pressure_thickness call, which is the
normal way to init dp3d in standalone tests. This was causing non-BFBness in the
C++-vs-F90 test for the bubble tests.

Differences in dp are at machine precision.