Change the KEY in output control from Casename to filename_prefix
It was decided that "Casename" is too abiguous for a KEY to control
what the output filename would be in IO.  A more descriptive key of
filename_prefix was preferred.  This commit makes those necessary
changes.

NOTE! All users will have to update their output control files
to use the new key.