And a helper function to scorpio_output to check field list for duplicates.
This commit adds a helper function `sort_and_check` which can be called
on the list of output fields and throw an error if the list has a duplicate
name.  The function also sorts the list of fields, which has no big
impact other than to put the list into alphabetical order.