Fix all instances of .get<Real> to be .get<double>
This commit fixes an anticipated bug if we move to support single
precision w/ dynamics, radiation and rayleigh_friction.  It changes
any use of the parameter list getter .get<Real> with .get<double>.

Note, although we cast the getter as double, because the variable
being defined is cast as Real we retain single precision.

Fixes #2234