EAMxx: do not mark double->float conversions as narrowing
Although double->float *does* have a loss of precision,
it is not a catastrophic one like float->int, where the
whole fractional part of the number is lost.

This allows tendency updates to work in both double and
single precision, since the timestep (used to form the
update coefficients) is always a double.