Add fuller description of algorithms to the module
There is no paper describing the algorithms in this module in any
detail, and the comments in the code are somewhat cryptic. A fuller
description of what is going on is added to the comments at the
beginning of routine shr_reprosum_calc and some comments are expanded
upon with the code.

BFB