isolate final copy to cpl data to its own do_export_for_cpl routine.
This commit split do_export one more time to isolate the step where
all of the coupler data is copied to the surface coupling data structures.

This commit also fleshes out the do_export_constant routine to actually copy
the constant value to the chosen fields.  NOTE: we still eventually overwrite
those copies in the "from_eamxx" export step.  The next commit will update the
do_export_from_eamxx routine to only update fields that are not a constant value.