Update stomatal resistance calculation when either rssun or rssha is 0
The changes follows Sam Silva's suggestion. Sam pointed out the change
really just serves as an error check. The physical meaning of rssha == 0
is very unclear; this mean that the stomatal conductance (rate of gas exchange)
is infinity. Infinitely large stomata minorly violates nearly every
conservation law in the known universe. It is suggested to set rs = 1e36
anytime either rssun or rssha is zero and fsun == 0 or 1, respectively.

[NBFB]