Establish a class to handle time interpolation between 2 fields of data.
This commit establishes the TimeInterpolation class as a utility that
can be used to
  a) store data for specific fields at two different timesnaps, and
  b) conduct a time interpolation between those two sets of data and
     produce an output.

The class stores two local field managers which in turn store the data
for two sets of data.  The timestamp associated with each set of data
is also stored so that a linear interpolation can be conducted on demand.

The data can be updated by passing a field of data to the class, which
will cause it to shift the data from time 1 to time 0 and replace the data
at time 1 with the new field of data.