Add infrastructure to handle interpolation data gathered from files.
This commit expands the infrastructure of TimeInterpolation to include
functions for dealing with 1+ files of interpolation data.

The files are sorted by timestamps to ensure that every interpolation step
the interpolation data used is from the time snaps closest to the target
time.

A subsequent commit will introduce a set of unit tests for this new
capability.