Add getter for variable metadata in a file, use it for time interpolation.
This commit adds a getter to the scream_scorpio interface that will allow
a user to grab metadata for a specific variable, rather than just the GLOBAL
metadata.

This is then used to get the units of time for a file passed to TimeInterpolation
so that the timestamps can be correctly set.