EAMxx: Add a tolerance on #diff lines in exxhash checker.
Flushed prints to e3sm.log can sometimes conflict with other output. Permit up
to 'thr' diffs so we don't fail due to badly printed lines. This isn't a big
loss in checking because an ERS_Ln22 second run writes > 1000 hash lines, and a
true loss of BFBness is nearly certain to propagate to a large number of
subsequent hashes.