Warn if debugTracers enabled with single layer in RK4
Recently, it had been made a critical error when debugTracers are
used with RK4 in single-layer mode.  This was because this
configuration is incompatible with many single-layer test cases.
However, we would also shown that the cosine-bell advection
test case does work with debugTracers and single layer.  To
support this case, we switch from a critical error to a warning
in this merge.