This commit constructs the surface coupling data in house.
This commit changes the surface coupling unit test to generate its
own set of prescribed export data rather than rely on a premade
netCDF file.  It also adds a check to the TimeInterpolation structure
to make sure we don't have multiple copies of the same timesnap of data
in our files.