Add nudging_fields as a runtime option in nudging.
This commit allows the user to pass the list of fields to be nudged as a
runtime option called `nudging_fields`.  In addition, this commit changes
the runtime option of `nudging_file` to be a vector of strings, allowing
users to pass more than one file.
NOTE: At the time of this commit only the first file is used, a subsequent
commit will use multiple files.

This commit is also incomplete, the time interpolation routine in nudging
still assumes that the user wants to nudge four variables.  This hasn't been
changed here because the next commit will transition to using the TimeInterpolation
class so all of that code will be removed.

Additionally, the infrastructure is not in place to let the user "add_field" with
just a name and grid.  So we are still hard-coding four fields to be update fields.

A different workflow will expand the infrastructure to facilitate add a field by name
alone, which will allow us to code the `set_grids` routine with a set of generic fields
rather than hard-code the ones there.