EAMxx: prefer HINTS to PATHS when looking for tpls
That's b/c HINTS is processed *before* system paths,
while PATHS is processed *after* them. Hence, on systems
that have a system-installed version, PATHS would always
pick up the system one. While this is likely irrelevant for
e3sm-shipped tpls like mct or csm_share, netcdf is sometimes
found in the system lib/inc dirs on some machines (and it
is usually not the right version).