Extend the set_variable_metadata function to accept Real valued input.
This commit extends the set_variable_metadata function in the IO interface
to allow users to pass a Real valued metadata value.  This is then used to
populate a "_FillValue" for each output set in the output.

The "_FillValue" name should match the convention applied in a typical EAM
simulation.