Support for optional alternative names in TimeInterpolation
This commit adds an option in TimeInterpolation to specify an alternative
name for a field when registering it with the TimeInterpolation structure.

The alternative name allows flexibility in using data from files where the
variable name in the file doesn't match the internal name in EAMxx.  The
alternative name will be the name assigned to the field internally, i.e. should
match the name in the file.  If no alternative name is provided than it is assummed
that the field name at registration matches the variable name in the file data.