Response to review: improve syntax used for alt names in surface coupling.
This commit improves the syntax used in the designation of alternative names in
the sc_export runtime options.  Now a user specifies which forced variables have
alternative names in the forcing file using a key pair with the pattern
  `EAMxx_variable_name:File_variable_name`
The special character `:` is used to separate the two tokens in the pair.