Add support in scorpio_output for tracking averaging count.
This commit adds a view to track the averaging count for every layout
represented in model output.  The intent of these views is to track
cases where filled values cause the averaging count to be spatially
varying.  This is essential for maintaing the appropriate average count
over restarts which use a restart history file.

Subsequent work will focus on setting and using these views properly, this
commit just focuses on establishing the views in the output variables.