Review Response: Revert combine function in scorpio_output.cpp
This commit reverts the original definition of the function `combine` in
scorpio_output.cpp and adds a new function specifically meant for cases
where filled values might be present, `combine_and_fill`.  This commit
also adds a bool variable that flags whether or not the output stream should
be tracking filled values.