Start to track the averaging count in scorpio output, and use it.
This commit introduces the infrastructure to start tracking which points
in a given layout are "filled" and thus shouldn't contribute to the
averaging total.  This is achieved by establishing a local set of views
per layout for each timestep, which records which points are filled for
this timesnap.  This is used to update both a) the overall count of timesnaps
that contribute to the average amount and b) the actual averaged amount.

Note, the averaging count is now included as an output.  This is specifically needed
for restart history files, which need to record this value to maintain BFB in
output w.r.t. restarts.  It is also helpful for users in identifying how often a value
was filled for a specific layout.  Note, that there is only one fill tracking variable
for each layout, so we don't expect the addition of these variables to significantly
impact the memory signature of our files.

Also note, there is a new logical which flags whether or not an output stream needs to
track filled values.  This defaults to FALSE but will be switched to TRUE if vertical
remapping is triggered or if the user specifies a fill value specifically for the output
stream.

A subsequent commit will clean up the code, remove redundant code and add the ability to
control if masking is tracked to the parameter list for an output stream.